requirements.txt
asgiref==3.4.1
Django==3.2.7
django-crispy-forms==1.13.0
django-js-asset==1.2.2
django-mptt==0.13.4
pytz==2021.1
sqlparse==0.4.2
